requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
reportlab==4.0.7
requests==2.31.0
supabase==2.3.4
python-dotenv==1.0.0
gunicorn==21.2.0

